project_4/requirements.txt
altair==5.3.0
ansi2html==1.9.2
asttokens==2.4.1
attrs==23.2.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colour==0.1.5
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
dash==2.17.1
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dash_auth==2.3.0
debugpy==1.8.2
decorator==5.1.1
dtreeviz==2.2.2
exceptiongroup==1.2.2
executing==2.0.1
explainerdashboard==0.4.7
Flask==3.0.3
Flask-WTF==1.2.1
flask_simplelogin==0.1.3
fonttools==4.53.1
gitdb==4.0.11
GitPython==3.1.43
graphviz==0.20.3
idna==3.7
imageio==2.34.2
importlib_metadata==8.0.0
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-dash==0.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
lazy_loader==0.4
lime==0.2.0.1
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
nest-asyncio==1.6.0
networkx==3.3
numba==0.60.0
numpy==2.0.0
oyaml==1.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
plotly==5.22.0
pluggy==1.5.0
prompt_toolkit==3.0.47
protobuf==5.27.2
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.2
pytest==8.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
requests==2.32.3
retrying==1.3.4
rich==13.7.1
rpds-py==0.19.0
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
shap==0.46.0
six==1.16.0
slicer==0.0.8
smmap==5.0.1
stack-data==0.6.3
streamlit==1.36.0
tenacity==8.5.0
threadpoolctl==3.5.0
tifffile==2024.7.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
waitress==3.0.0
watchdog==4.0.1
wcwidth==0.2.13
Werkzeug==3.0.3
WTForms==3.1.2
zipp==3.19.2
